+++ plugins/jsvm/binds_test.go
@@ -998,7 +998,7 @@ func TestApisBindsCount(t *testing.T) {
 	apisBinds(vm)
 
 	testBindsCount(vm, "this", 8, t)
-	testBindsCount(vm, "$apis", 12, t)
+	testBindsCount(vm, "$apis", 11, t)
 }
 
 func TestApisBindsApiError(t *testing.T) {